refatoração

+++ aluno.h
@@ -1,5 +1,5 @@
 typedef struct aluno{
-    int CPF; // Pra hash só usamos os 9 primeiros
+    long long CPF; // Pra hash só usamos os 9 primeiros
     char* nome;
     int nota;
 } Taluno;
